Renames: top-layout -> main-layout, top-content -> main-content
