+++ src/components/main-layout/main-content.less
@@ -1,6 +1,6 @@
 @import '../../styles/metrics';
 
-.top-content {
+.main-content {
 	min-height: 100vh;
 	padding: calc(50px + 2 * @len-large) @len-large @len-large @len-large;
 }
