+++ src/routes/passage-edit/index.less
@@ -1,5 +1,5 @@
 .passage-edit {
-	.top-content {
+	.main-content {
 		display: flex;
 		flex-direction: column;
