+++ src/routes/story-edit-javascript/index.less
@@ -1,5 +1,5 @@
 .story-edit-javascript {
-	.top-content {
+	.main-content {
 		display: flex;
 		flex-direction: column;
