+++ src/routes/story-edit-stylesheet/index.less
@@ -1,5 +1,5 @@
 .story-edit-stylesheet {
-	.top-content {
+	.main-content {
 		display: flex;
 		flex-direction: column;
