+++ src/routes/story-edit-stylesheet/index.vue
@@ -4,21 +4,21 @@
 			:back-route="`/stories/${this.story.id}`"
 			:back-label="story.name"
 		/>
-		<top-content :title="$t('editors.storyStylesheet.dialogTitle')">
+		<main-content :title="$t('editors.storyStylesheet.dialogTitle')">
 			<p v-t="'editors.storyStylesheet.dialogExplanation'" />
 			<code-area @change="onChange" mode="css" :value="story.stylesheet" />
-		</top-content>
+		</main-content>
 	</div>
 </template>
 
 <script>
 import CodeArea from '@/components/input/code-area';
-import TopBar from '@/components/top-layout/top-bar';
-import TopContent from '@/components/top-layout/top-content';
+import TopBar from '@/components/main-layout/top-bar';
+import MainContent from '@/components/main-layout/main-content';
 import './index.less';
 
 export default {
-	components: {CodeArea, TopBar, TopContent},
+	components: {CodeArea, MainContent, TopBar},
 	computed: {
 		story() {
 			const result = this.$store.state.story.stories.find(
